User/requirements.txt
bcrypt==3.2.0
cffi==1.14.4
grpcio==1.34.0
grpcio-tools==1.34.0
protobuf==3.14.0
pycparser==2.20
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.20
tornado==6.1
